test(manual-tests): v0.3.15 arc COMPLETE — all ten legs live; every scoped issue evidenced
Legs 5-9 executed against main tip 9fccb6f5, completing the arc.
$0.1343 across 29 reconciled leases, measured before teardown.

RESULTS

Leg 5 (the travel): 28 lines of autonomous roundtable transcript, ZERO
hits for Mira / surgery / review slot / alice. The content sits under
alice-person and agent-origin turns resolve local, so it cannot be
reached — an absence bar with real transcript to be absent from.

Leg 6 (close-trigger asymmetry): an Alice-origin close produced the same
shape as Leg 4's idle close. The record binds its own frozen principal,
so the trigger no longer selects a tenant. This leg was load-bearing
because Leg 3's quorum never formed.

Leg 7: alice-person, bob-person and local coexist as separate partitions
in every store — isolation is a recall filter, not a delete. But the
absence bar itself passed VACUOUSLY: no persona replied to Bob at all
(F-5), which is Edge Case 1's own trap.

Leg 8: the same sender alice-person is stamped principal='alice-person'
under auth and 'local' with it off — sender_id and principal are
genuinely separate axes. The local partition holds TWO speakers, not
one: the speaker split is auth-independent, as the leg asserts.
ISSUE-0125 proved more strongly than asked — THREE orchestrator
restarts, 6/6 agents registered, every agent container Up 39 minutes
with no process touched.

Leg 9 (ISSUE-0130(b)): A -> B grows (0 -> 7/9/7 replay rows), and
B = C byte-identical in every partition. Replay derives under the
ATTRIBUTED tenant — replay- rows spread across alice-person, bob-person
and local, preserving the speaker axis — not into local wholesale. The
discriminating follow-up (one message, third restart) added exactly +1
per persona, which is what separates "the guard bounds duplication" from
"replay stopped deriving entirely".

THREE MORE MT DEFECTS (ten total; v1.1 had never been executed)

8. Leg 5's convene is wrong twice: roundtable ships DISARMED, and
   convene takes no topic argument. Arm at runtime first.
9. Leg 9 needs alice, but Leg 7 destroyed her — account bootstrap is
   single-account, so the rotation to bob REPLACES alice. "Re-
   authenticate as alice" fails invalid credentials against an account
   that no longer exists. Rotate back; that restart also closes what
   Leg 8 left open, which is how the local partition gets its records.
10. Leg 7's probe does not compel a reply (F-5). Recorded for the next
    run rather than fixed blind — it needs a live run to confirm a
    mention actually draws one.

DRIVER FIXES

- bootstrap_account now performs the orchestrator restart itself. It is
  not the caller's job: the orchestrator holds the deleted accounts.db
  inode until it reopens, so a login before it fails invalid credentials
  against an account that plainly exists on disk. Legs 0 and 7 called it
  separately and Leg 9 forgot, losing a run.
- Artifacts are now written on the abort path too. The first Leg 9
  failure discarded four completed legs' evidence already paid for.
- Legs 5, 8 and 9 no longer print ACTION REQUIRED: arming, the auth flip
  and the rotation are automated, so the arc is unattended.
- set_auth_mode rewrites only the setting line inside the auth block — a
  naive substitution would hit the comment that made F-1 possible.

Driver crossed the 500-line cap, so the operations vocabulary split to
mt_group_tenant_ops.py (350 + 247). The seam is verbs vs the arc.

Every scoped issue is now evidenced: ISSUE-0123, 0124, 0130(b), 0131,
0125/0126. Outstanding before this PR can close: MT-MEMORY-MULTIUSER-001
re-run, eval-replay, the offline smoke, and reverting the run knobs.
