+++ FILEMAP.md
@@ -1,7 +1,7 @@
 # Repository File Map
 
 > Auto-generated by `scripts/generate_filemap.py` — 2026-09-02  
-> **2069** tracked files
+> **2070** tracked files
 
 ## Summary
 
@@ -14,7 +14,7 @@
 | `web/` | 78 |
 | `prompts/` | 48 |
 | `cli/` | 42 |
-| `scripts/` | 31 |
+| `scripts/` | 32 |
 | `(root)` | 30 |
 | `cmd/` | 23 |
 | `config/` | 22 |
@@ -28,7 +28,7 @@
 
 | Extension | Count |
 |-----------|-------|
-| `.py` | 783 |
+| `.py` | 784 |
 | `.go` | 542 |
 | `.md` | 528 |
 | `.yaml` | 57 |
@@ -1679,6 +1679,7 @@
 │   │   ├── __init__.py
 │   │   ├── mt_group_tenant_001.py
 │   │   ├── mt_group_tenant_evidence.py
+│   │   ├── mt_group_tenant_ops.py
 │   │   └── mt_group_tenant_preflight.py
 │   ├── perf/
 │   │   └── wallet_p99.py
